serverless/requirements.txt
runpod>=1.7.0
pillow>=10.4.0
